+++ age3-core/src/main/java/pl/edu/agh/age/client/WorkerServiceClient.java
@@ -36,6 +36,8 @@ public interface WorkerServiceClient {
 
 	boolean isComputationRunning();
 
+	boolean isComputationFinished();
+
 	boolean isComputationFailed();
 
 	ComputationState computationState();
@@ -43,4 +45,6 @@ public interface WorkerServiceClient {
 	Optional<WorkerConfiguration> currentConfiguration();
 
 	Optional<Throwable> currentError();
+
+	void waitForComputationEnd() throws InterruptedException;
 }