+++ age3-core/src/main/java/pl/edu/agh/age/services/worker/internal/CommunicationFacility.java
@@ -48,4 +48,6 @@ public interface CommunicationFacility {
 	Set<WorkerMessage.Type> subscribedTypes();
 
 	void start();
+
+	void reset();
 }