+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 Console:
 - Add the batch mode to the console (#25).
 - `pl.edu.agh.age.console.command.ComputationCommand.info` provides now real info about computation (#26).
+- Add `computation.waitUntilFinished` command (#34).
 
 Core:
 - Compute setup failures do not result in node failure anymore (#24).
@@ -13,10 +14,12 @@ Core:
 General compute:
 - Introduce topology API (`pl.edu.agh.age.compute.api.TopologyProvider`) for compute-level applications
   and use it in StreamAgents migrations (#30).
+- Full clean up is performed on all utilities during task clean-up (#34).
 
 Stream Agents:
 - Use topology API in migrations (#30).
 - Make `Improvement` run after evaluation of the solution when configured using `SexualReproductionBuilder` (#33).
+- Fix situations when the main runnable did not wait for workplaces (#29). 
 
 ## 0.3.1 -- TBA